# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel TechTest API",
        "description": "Laravel TechTest API description",
        "contact": {
            "email": "keoz.mm@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/company/all": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display a listing of the resource.",
                "operationId": "company@index",
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/company/create": {
            "post": {
                "tags": [
                    "company"
                ],
                "summary": "Store a newly created resource in storage.",
                "operationId": "company@store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Company name",
                                        "type": "string",
                                        "example": "Prova Contratto S.A de C.V"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "hello@provacontratto.com.sv"
                                    },
                                    "phone": {
                                        "description": "Phone",
                                        "type": "string",
                                        "example": "(503) 2558-8798"
                                    },
                                    "city": {
                                        "description": "City",
                                        "type": "string",
                                        "example": "San Vicente"
                                    },
                                    "country": {
                                        "description": "Country",
                                        "type": "string",
                                        "example": "El Salvador"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet consectetur adipisicing elit."
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "example": "Activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/company/filter/id/{id}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display the specified resource.",
                "operationId": "company@showById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company to get by id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/company/filter/name/{name}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display the specified resource LIKE Name.",
                "operationId": "company@showByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Company to get by name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/company/filter/email/{email}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display the specified resource LIKE Email.",
                "operationId": "company@showByEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Company to get by email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/company/filter/phone/{phone}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display the specified resource LIKE Phone.",
                "operationId": "company@showByPhone",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "Company to get by phone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/company/filter/city/{city}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display the specified resource LIKE City.",
                "operationId": "company@showByCity",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "Company to get by city",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/company/filter/country/{country}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display the specified resource LIKE Country.",
                "operationId": "company@showByCountry",
                "parameters": [
                    {
                        "name": "country",
                        "in": "path",
                        "description": "Company to get by country",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/company/filter/address/{address}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display the specified resource LIKE Address.",
                "operationId": "company@showByAddress",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Company to get by address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/company/filter/status/{status}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Display the specified resource by Status.",
                "operationId": "company@showByStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Company to get by status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
...
        "/api/company/delete/{id}": {
            "delete": {
                "tags": [
                    "company"
                ],
                "summary": "Remove the specified resource from storage.",
                "operationId": "company@destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company to delete by id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
...
    }
}